+++ src/common/CMakeLists.txt
@@ -24,20 +24,20 @@ add_library(common
             metrics.cpp
             metrics.hpp
 	    png_reader.hpp
+            png_reader.cpp
             )
 
 target_include_directories(common
     PUBLIC .
     PRIVATE SYSTEM ${CONAN_INCLUDE_DIRS_JSONFORMODERNCPP}
     PRIVATE SYSTEM ${CONAN_INCLUDE_DIRS_CLI11}
-    PRIVATE SYSTEM ${CONAN_INCLUDE_DIRS_LIBPNG}
+    PUBLIC SYSTEM ${CONAN_INCLUDE_DIRS_LIBPNG}
     PUBLIC SYSTEM ${CONAN_INCLUDE_DIRS_SPDLOG}
 )
 
 target_link_libraries(common
   PUBLIC ${CONAN_LIBS_FMT}
   PUBLIC ${CONAN_LIBS_LIBPNG}
-  PRIVATE ${CONAN_LIBS_ZLIB}
 )
 
 target_compile_definitions(common PRIVATE ${SHAR_COMPILE_DEFINITIONS})
